Crate gooey_kludgine

A Renderer for Gooey that uses Kludgine to draw. Under the hood, Kludgine uses wgpu, and in the future we aim to support embedding Gooey into other wgpu applications.

User interface scaling (Points)

Kludgine uses winit to determine the scaling ratio to use. For more information on the approaches taken, see the winit::dpi module.

Re-exports

pub use kludgine;

Structs

Kludgine